+++ jason/src/peer/media/mod.rs
@@ -750,6 +750,15 @@ impl MediaConnections {
         );
         remove_tracks_fut.await;
     }
+
+    /// Removes a [`sender::Component`] or a [`receiver::Component`] with the
+    /// provided [`TrackId`] from these [`MediaConnections`].
+    pub fn remove_track(&self, track_id: TrackId) {
+        let mut inner = self.0.borrow_mut();
+        if inner.receivers.remove(&track_id).is_none() {
+            inner.senders.remove(&track_id);
+        }
+    }
 }
 
 #[cfg(feature = "mockable")]